# apps/otoge-bookmark-api/src/games/games.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  Delete,
  UseGuards,
  NotFoundException,
  Query,
  BadRequestException,
} from '@nestjs/common';
import {
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiOkResponse,
  ApiTags,
} from '@nestjs/swagger';
import {
  DEFAULT_TAKE,
  DEFAULT_ORDER_BY,
} from '@/common/constants/list.constant';
import { JwtAuthGuard } from '@/common/guards/jwt-auth.guard';
import { CreateGameDto } from './dto/create-game.dto';
import { ListGameDto } from './dto/list-game.dto';
import { UpdateGameDto } from './dto/update-game.dto';
import { GameEntity, GameWithSongsEntity } from './entities/game.entity';
import { GamesService } from './games.service';

@Controller('games')
@ApiTags('games')
export class GamesController {
  constructor(private readonly gamesService: GamesService) {}

  @Post()
  @UseGuards(JwtAuthGuard)
  @ApiCreatedResponse({ type: GameEntity })
  async create(@Body() data: CreateGameDto) {
    return new GameEntity(await this.gamesService.create(data));
  }

  @Get()
  @ApiOkResponse({ type: GameWithSongsEntity, isArray: true })
  async findAll(@Query() query: ListGameDto) {
    const { skip, take, cursor, orderBy } = query;

    if (skip && isNaN(Number(skip))) {
      throw new BadRequestException('skip must be number');
    }

    if (take && isNaN(Number(take))) {
      throw new BadRequestException('take must be number');
    }

    if (cursor) {
      const game = await this.gamesService.findOne({ id: cursor });
      if (!game) {
        throw new NotFoundException('cursor is invalid');
      }
    }

    const games = await this.gamesService.findAll({
      skip: skip ? Number(skip) : 0,
      take: take ? Number(take) : DEFAULT_TAKE,
      cursor: cursor
        ? {
            id: cursor,
          }
        : undefined,
      orderBy: {
        createdAt: orderBy ?? DEFAULT_ORDER_BY,
      },
    });

    if (!games.length) {
      throw new NotFoundException();
    }

    return games.map((game) => {
      return new GameWithSongsEntity(game);
    });
  }

  @Get(':id')
  @UseGuards(JwtAuthGuard)
  @ApiOkResponse({ type: GameWithSongsEntity })
  async findOne(@Param('id') id: string) {
    const game = await this.gamesService.findOne({ id });
    if (!game) {
      throw new NotFoundException();
    }
    return new GameWithSongsEntity(game);
  }

  @Patch(':id')
  @UseGuards(JwtAuthGuard)
  @ApiOkResponse({ type: GameEntity })
  async update(@Param('id') id: string, @Body() data: UpdateGameDto) {
    return new GameEntity(
      await this.gamesService.update({ where: { id }, data }),
    );
  }

  @UseGuards(JwtAuthGuard)
  @Delete(':id')
  @ApiNoContentResponse()
  async remove(@Param('id') id: string): Promise<void> {
    const game = await this.gamesService.findOne({ id });
    if (!game) {
      throw new NotFoundException();
    }

    this.gamesService.remove({ id });
  }
}